% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Data}
\alias{Data}
\title{Telemetry data of red deer}
\format{
A data frame with 1109 rows and 4 variables:
\describe{
\item{x}{x-coordinate}
\item{y}{y-coordinate}
\item{burst}{name of id}
\item{date}{time of recording}
}
}
\source{
{Bavarian Forest Nationalpark}
}
\usage{
Data
}
\description{
A dataset containing the coordinates (x,y), id (burst) and time of recording (date).
}
\keyword{datasets}
