% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_avail.R
\name{get_avail}
\alias{get_avail}
\title{Availability of habitats within a home range}
\usage{
get_avail(PP, Landscapegrid, plot = FALSE)
}
\arguments{
\item{PP}{data.frame than contains x- and y- coordinates, the name of the individual as $burst and if a plot is wished the time as $ti}

\item{Landscapegrid}{SpatialGridDataFrame where names of the habitat types are stored as names "habitat" in the data}

\item{plot}{(logical) wether you wish to see a plot of the home range within the landscape (time comsuming)}
}
\description{
This function allows you to obtain the number of grid cells covered with each habitat type within a 100\% mcp
home range spanned by given x- and y- coordinates. A SpatialGridDataFrame that provides landscape information is needed.
It is advicable to generate the data.frame PP via ld(as.ltraj(data)).
This function is used within \code{\link{get_avail_indi_ti}}.
}
\examples{
...
}
\keyword{home}
\keyword{ranges}
