% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Holling1.R
\name{Holling1}
\alias{Holling1}
\title{Function for Holling's type I}
\usage{
Holling1(x, a)
}
\arguments{
\item{x}{value}

\item{a}{parameter for Holling type I}
}
\value{

}
\description{
Function for Holling's type I
}
\examples{

}
\keyword{linear}
