% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRti.R
\name{HRti}
\alias{HRti}
\title{Availability of habitats within time-indexed home ranges}
\usage{
HRti(Df, varti, Landscapegrid, discN = 0, plot = FALSE)
}
\arguments{
\item{Df}{data.frame than contains x- and y- coordinates, the name of the individual as $burst and the times ti, over which levels the home ranges are calculated}

\item{varti}{column name of Df that encodes the time intervals. It can be continous, but is set to floor to ensure categorical time intervals}

\item{Landscapegrid}{SpatialGridDataFrame where names of the habitat types are stored as names "habitat" in the data}

\item{discN}{takes only time intervals ti with more than discN recordings per time interval ti}

\item{plot}{(logical) wether you wish to see a plot of the home range within the landscape (time comsuming)}
}
\value{
data.frame, each row refers to one individual and one time interval,
columns are
"burst",
"ti"(time interval),
"total" (number of grid cells that span the home range),
and for each habitat type one column that contains the number of grid cells overlayed by the habitat#'
}
\description{
This function allows you to obtain the a list of the number of grid cells covered with each habitat type within a 100\% mcp
home ranges spanned by given x- and y- coordinates over well defined time intervals for a group of individuals.
A SpatialGridDataFrame that provides landscape information is needed.
It is advicable to generate the data.frame Df via adehabitatHR::ld(as.ltraj(data)).
}
\examples{
data(Data)
data(Landscapegrid) #Landscapegrid needs column in datamatrix called "habitat"

Data$month <- as.numeric(format(Data$date,"\%m"))
Data$hour <- as.numeric(format(Data$date,"\%H"))
table(Data[,c("month","burst")])
get landscape information
Data$habitat  <- over(SpatialPoints(cbind(Data$x,Data$y)),Landscapegrid)$habitat
Data$habitat  <- relevel(Data$habitat, ref="old mixed")

#get availability of habitat types in monthly home range
par(mfrow=c(3,4),mar=c(1,1,1,1),oma=c(1,1,1,1))
Df$ti <- floor(Df[,"month"])		
#get availability of habitats within time-indexed home ranges
TTihm <- table(Df[,c("burst","ti")])

get availability of habitat types in monthly home range
par(mfrow=c(3,4),mar=c(1,1,1,1),oma=c(1,1,1,1))
HRavail <- HRti(Data,var="month",Landscapegrid,plot=TRUE)
}
\keyword{home}
\keyword{ranges}
